fix: remove cyclic dep
