+++ packages/fuel-indexer-macros/Cargo.toml
@@ -30,12 +30,5 @@ serde_json = { workspace = true }
 sha2 = "0.10"
 syn = { version = "2.0", features = ["full"] }
 
-[dev-dependencies]
-fuel-indexer-plugin = { workspace = true }
-fuel-indexer-utils = { workspace = true }
-fuels-macros = { version = "0.46", default-features = false }
-serde = { workspace = true }
-trybuild = "1.0"
-
 [features]
 default = ["fuel-indexer-schema/db-models"]
