+++ packages/fuel-indexer-tests/tests/lib.rs
@@ -1,10 +1,14 @@
-#[cfg(feature = "postgres")]
+#[cfg(all(feature = "postgres", not(feature = "trybuild")))]
 mod graphql_server;
 
-#[cfg(feature = "postgres")]
+#[cfg(all(feature = "postgres", not(feature = "trybuild")))]
 mod indexing;
 
+#[cfg(not(feature = "trybuild"))]
 mod service;
 
-#[cfg(feature = "postgres")]
+#[cfg(all(feature = "postgres", not(feature = "trybuild")))]
 mod web_server;
+
+#[cfg(feature = "trybuild")]
+mod trybuild;
