+++ packages/fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.stderr
@@ -25,15 +25,16 @@ error[E0433]: failed to resolve: use of undeclared type `Consensus`
   |                    ^^^^^^^^^ use of undeclared type `Consensus`
   |
 help: consider importing one of these items
-   |
+  |
 2  + use crate::fuel::Consensus;
+   |
+2  + use fuel_indexer::prelude::fuel::Consensus;
    |
 2  + use fuel_indexer_plugin::prelude::fuel::Consensus;
    |
 2  + use fuel_indexer_types::fuel::Consensus;
    |
-2  + use fuel_indexer_utils::plugin::prelude::fuel::Consensus;
-   |
+     and 1 other candidate
 
 error[E0433]: failed to resolve: use of undeclared type `Transaction`
  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
@@ -42,41 +43,41 @@ error[E0433]: failed to resolve: use of undeclared type `Transaction`
   |                          ^^^^^^^^^^^ use of undeclared type `Transaction`
   |
 help: consider importing one of these items
-   |
+  |
 2  + use crate::fuel::Transaction;
+   |
+2  + use fuel_indexer::prelude::fuel::Transaction;
    |
 2  + use fuel_indexer_plugin::prelude::fuel::Transaction;
    |
 2  + use fuel_indexer_types::fuel::Transaction;
    |
-2  + use fuel_indexer_utils::plugin::prelude::fuel::Transaction;
-   |
-     and 1 other candidate
+     and 4 other candidates
 
-error[E0560]: struct `Header` has no field named `output_messages_count`
+error[E0560]: struct `fuel_indexer_utils::prelude::fuel::Header` has no field named `output_messages_count`
  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
   |
   |             output_messages_count: 1,
-  |             ^^^^^^^^^^^^^^^^^^^^^ `Header` does not have this field
+  |             ^^^^^^^^^^^^^^^^^^^^^ `fuel_indexer_utils::prelude::fuel::Header` does not have this field
   |
   = note: available fields are: `id`, `da_height`, `transactions_count`, `message_receipt_count`, `transactions_root` ... and 5 others
 
-error[E0560]: struct `Header` has no field named `output_messages_root`
+error[E0560]: struct `fuel_indexer_utils::prelude::fuel::Header` has no field named `output_messages_root`
  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
   |
   |             output_messages_root: [0u8; 32].into(),
-  |             ^^^^^^^^^^^^^^^^^^^^ `Header` does not have this field
+  |             ^^^^^^^^^^^^^^^^^^^^ `fuel_indexer_utils::prelude::fuel::Header` does not have this field
   |
   = note: available fields are: `id`, `da_height`, `transactions_count`, `message_receipt_count`, `transactions_root` ... and 5 others
 
 error[E0308]: mismatched types
- --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
-  |
-  |                     data: bytes,
-  |                           ^^^^^ expected `Option<Vec<u8>>`, found `Vec<u8>`
-  |
-  = note: expected enum `std::option::Option<Vec<_>>`
-           found struct `Vec<_>`
+  --> ../fuel-indexer-tests/trybuild/fail_if_attribute_schema_arg_is_invalid.rs
+   |
+   |                     data: bytes,
+   |                           ^^^^^ expected `Option<Vec<u8>>`, found `Vec<u8>`
+   |
+   = note: expected enum `std::option::Option<Vec<_>>`
+            found struct `Vec<_>`
 help: try wrapping the expression in `Some`
    |
 65 |                     data: Some(bytes),
